♻️ 장기 게임 구현에 필요한 상수들 constants.ts에 보관
